Recognize when HSMP is not enabled
BIOS on systems provides a method to disable the HSMP mailboxes.
Doing this will result in HAMP requests to time out. As of now
this is the only indicator we have that HSMP is disabled.

Update libhsmp to recognize when a timeout occurs for a HSMP call
and set a flag to indicate that HSMP is disabled. This allows
the library to short circuit any future calls so it does go
through the set up process and simply returns ENOTSUP.

Additionally, to properly handle this major updates had to be
made to the hsmp_test test program.

Signed-off-by: Nathan Fontenot <nathan.fontenot@amd.com>