Fix use of socket ID vs. NBIO index
The intent of the nbios array stored in hsmp_data is to hold the NBIO ID
(unique to each socket) and the PCI bus base and limit values for each
NBIO tile to assist with translating a PCI bus to the socket and NBIO
tile ID hosting that bus. The PCI device for each NBIO is kept here as
well as a matter of convenience. All HSMP messages must be sent to an
IOHC device in the specific socket (0 or 1) for which the message is
intended. After sorting the NBIOs array based on the bus base/limit,
index 0 into that array will correspond with the lowest bus in socket 0,
and if present, index 4 into that array will correspond with the lowest
bus in socket 1. While technically using any IOHC device in a socket
will work to communicate with the HSMP mailbox, the first (lowest bus
number) IOHC device is used as a matter of convention. Therefore, when
accessing the correct PCI device, socket ID must be multiplied by 4 to
get to the NBIO array index.

In a 1P system, NBIO array indexes 4-7 will be empty, so test for a
NULL pointer for the PCI device.

Signed-off-by: Lewis Carroll <lewis.carroll@amd.com>
Signed-off-by: Nathan Fontenot <nathan.fontenot@amd.com>