Set NBIO P-state fn should use bus number
The hsmp_set_nbio_pstate function should use the PCI bus number instead
of socket ID since there are four NBIO tiles per socket. Since the
end-user or software "sees" the PCI system as bus numbers, not as NBIO
tiles or root complex IDs, take the bus number and translate it to the
corresponding socket ID and NBIO tile ID to send the appropriate HSMP
message.

Update hsmp_test.c based on this change by removing the test for an
invalid bus number. Now that the bus number is specified as an u8
we don't have a invalid value that works. This should be re-added
if we can find a way to specify an invalid bus number.

Signed-off-by: Lewis Carroll <lewis.carroll@amd.com>
Signed-off-by: Nathan Fontenot <nathan.fontenot@amd.com>