Fix bug in hsmp_set_data_fabric_pstate
The enum for fabric P-state is 0-4 in the header file, with value 0
being the value to enable auto fabric P-state selection. This leaves
values 1 - 4 for fixed fabric P-state selection. However, in hardware
the P-states are numbered 0 - 3. Therefore, when setting fixed fabric
P-state we need to subtract 1 to account for the use of P-state 0 as the
value for auto P-state selection.

Signed-off-by: Lewis Carroll <lewis.carroll@amd.com>
Signed-off-by: Nathan Fontenot <nathan.fontenot@amd.com>