Add socket id parameter to fabric clock APIs
The initial implementation of the fabric clock APIs, hsmp_memory_clock()
and hsmp_data_fabric_clock(), did not include a socket is parameter.
This is incorrect as the values reported are a per socket value and not
a system value. This updates the clock PIs to include a socket id
parameter instead of defaulting to socket 0.

This update also adds a test case for an incorrect socket id to the
hsmp test suite.

Suggested-by: Lewis Carroll <lewis_carroll@amd.com>
Signed-off-by: Nathan Fontenot <nathan.fontenot@amd.com>