Correct DDR APIs to add socket_id
The initial PPR stated that the reported DDR bandwidth values were
reported for the system. This has been corrected as they are reported
on a per socket basis.

To support this the DDR APIs have a socket_id value added instead of
defaulting to socket 0 inside of the library.

This also adds a test case in the hsmp test suite for invalid socket
ids.

Suggested-by: Lewis Carroll <lewis.carroll@amd.com>
Signed-off-by: Nathan Fontenot <nathan.fontenot@amd.com>