Add socket to hsmp_data_fabric_pstate()
The initial implementation mistakenly did not provide a parameter to
specify the socket id when setting the data fabric pstate. The HSMP
interface only allows to set thios on a per-socket basis, this
update adds a socket_id parameter to the call.

This also adds a hsmnp test case for invalid socket is.

Suggested-by: Lewis Carroll <lewis.carroll@amd.com>
Signed-off-by: Nathan Fontenot <nathan.fontenot@amd.com>