Reorder DF P-state enum for code clarity
Starting enum hsmp_df_pstate with the value for auto means that the
values 1-4 result for P-states 0-3. The code has to subtract one before
passing this value to the hardware.

By moving the auto value to the end of the enum, the enum values are
aligned with the hardware values making the code a bit cleaner.

Signed-off-by: Lewis Carroll <lewis.carroll@amd.com>
Signed-off-by: Nathan Fontenot <nathan.fontenot@amd.com>