Update errno return values
To clarify the different failures in libhsmp this updates the
errno values the libhsmp library sets and ensures they are
consistently applied.

EPERM - set for any non-root caller to libhsmp.

ENOTSUP - set if HSMP is disabled in BIOS or the CPU mopdel/family
is not supported.

ENODEV - set when initialization of PCI bus enumeration in
libhsmp fails.

ENOMSG - set if the HSMP message id is not supported by the current
hsmp interface version.

EBADMSG - set if the HSMP message id is supported by the current
HSMP interface version but not supported by the SMU fw.

EAGAIN - set if a failure occurs during libhsmp initialization
where a future call to libhsmp may succeed.

EINVAL - set for any invalid parameter passed to libhsmp.

ETIMEDOUT - set when a call to the HSMP mailboxes times out

Suggested-by: Lewis Carroll <lewis.carroll@amd.com>
Signed-off-by: Nathan Fontenot <nathan.fontenot@amd.com>