libhsmp: Correct HSMP interface version support
Libhsmp HSMP interface version support is broken in that it
will assume support only for the highest interface version
the library currently supports. Currently this is version 3.
Using libhsmp on a system where the SMU supports a higher
interface version will result in libhsmp reporting that all
interfaces are unsupported. When on a system with a lower
SMU interface version libhsmp will still report all interfaces
supported by the library as supported instead of just those
supported by SMU.

To correct this a new interface version supported value is
used to check for interface support instead of using the
highest interface version supported by the library. This
value will be the min of the SMU reported interface version
and the highest interface version supported by the library.

Signed-off-by: Nathan Fontenot <nathan.fontenot@amd.com>