Report EBADMSG error results
Libhsmp reports an errno of EBADMSG when a called interface gets a
return code of not supported from SMU for the interface but the
HSMP interface version indicates support for the called interface.

This singling out of specific interfaces by different SMU versions
to not support a specific interface even though the HSMP interface
version SMU reports indicates support for the interface makes it
difficult to know if the call failed or properly received the
EBADMSG errno.

To help with testing, track return codes of EBADMSG separately
from pass/fail tests. Then at the end of the report list the
interfaces that received EBADMSG. This allow testing to easily
determine if the received EBADMSG was correct for the SMU
version.

Signed-off-by: Nathan Fontenot <nathan.fontenot@amd.com>