Replace u* types with stdint uint*_t types
The u* types used in libhsmp are made possible by the inclusion
of pci/types.h. Updates to the libhsmp code to move the nbio,
and thus the pci handling, code to a separate file we will no
longer include pci/types.h in libhsmp.c

To do this and make a more standard definition of types, update
libhsmp to use the standard uint*_t types instead of the u*
types defined in pci/types.h.

Signed-off-by: Nathan Fontenot <nathan.fontenot@amd.com>