Allocate NBIO Array
Current nbio initialization makes a few assumptions that do
not hold across current systems and may break in the future.
Namely this is that the maimum number of nbio devices is 8,
the maximum number of sockets is 2, and the number of nbios
per socket is 4.

Update the initialization of nbios to remove these assumptions
by allocating the array nbio devices based on the number of
discovered at init time. Set the number of sockets based on
what we discover during init and then use this to calculate
the number of nbios per socket.

Signed-off-by: Nathan Fontenot <nathan.fontenot@amd.com>